WIP event loop/queue

+++ src/grid.c
@@ -17,7 +17,7 @@ void grid_initialize(Grid *g)
 		{GRID_HALF_SIZE - 1, GRID_HALF_SIZE - 1},
 	};
 	quad_to_size_quad(&g->size_quad, &g->styles_quad);
-	int capacity = size_quad_capacity(&g->size_quad);
+	i64 capacity = size_quad_capacity(&g->size_quad);
 	g->styles = calloc(capacity, sizeof *g->styles);
 }
 
@@ -26,11 +26,11 @@ void grid_terminate(Grid *g)
 	free(g->styles);
 }
 
-static int grid_index(Grid *g, int2 h)
+static i64 grid_index(Grid *g, int2 h)
 {
 	assert(quad_contains(&g->quad, h));
-	int diff_min_x = h.x - g->size_quad.min.x;
-	int diff_min_y = h.y - g->size_quad.min.y;
+	i64 diff_min_x = h.x - g->size_quad.min.x;
+	i64 diff_min_y = h.y - g->size_quad.min.y;
 	return diff_min_x + diff_min_y * g->size_quad.size.x;
 }
 
@@ -43,13 +43,13 @@ u8 grid_get(Grid *g, int2 h)
 void grid_set(Grid *g, int2 h, u8 style)
 {
 	assert(quad_contains(&g->quad, h));
-	int i = grid_index(g, h);
+	i64 i = grid_index(g, h);
 	g->styles[i] = style;
 }
 
 void grid_clear(Grid *g, int2 h)
 {
 	assert(quad_contains(&g->quad, h));
-	int i = grid_index(g, h);
+	i64 i = grid_index(g, h);
 	g->styles[i] = 0;
 }
